Move FSpot.Widgets -> Widgets
