+++ src/Clients/FSpot/FSpot.csproj
@@ -40,7 +40,6 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="DependentListStore.cs" />
-    <Compile Include="FSpot.Widgets\TagSelectionWidget.cs" />
     <Compile Include="FSpot\App.cs" />
     <Compile Include="FSpot.ColorAdjustment\Adjustment.cs" />
     <Compile Include="FSpot.ColorAdjustment\Desaturate.cs" />
@@ -81,7 +80,6 @@
     <Compile Include="FSpot.Filters\SharpFilter.cs" />
     <Compile Include="FSpot.Filters\UniqueNameFilter.cs" />
     <Compile Include="FSpot.Filters\WhiteListFilter.cs" />
-    <Compile Include="FSpot\FolderQueryWidget.cs" />
     <Compile Include="FSpot\FullScreenView.cs" />
     <Compile Include="FSpot\GroupAdaptor.cs" />
     <Compile Include="FSpot\GroupSelector.cs" />
@@ -126,30 +124,47 @@
     <Compile Include="FSpot.UI.Dialog\TagSelectionDialog.cs" />
     <Compile Include="FSpot.UI.Dialog\ThreadProgressDialog.cs" />
     <Compile Include="FSpot\UriCollection.cs" />
-    <Compile Include="FSpot.Widgets\CellRendererTextProgress.cs" />
-    <Compile Include="FSpot.Widgets\EditorPage.cs" />
-    <Compile Include="FSpot.Widgets\Filmstrip.cs" />
-    <Compile Include="FSpot.Widgets\FindBar.cs" />
-    <Compile Include="FSpot.Widgets\FolderTreeModel.cs" />
-    <Compile Include="FSpot.Widgets\FolderTreePage.cs" />
-    <Compile Include="FSpot.Widgets\FolderTreeView.cs" />
-    <Compile Include="FSpot.Widgets\ImageInfo.cs" />
-    <Compile Include="FSpot.Widgets\InfoBox.cs" />
-    <Compile Include="FSpot.Widgets\Loupe.cs" />
-    <Compile Include="FSpot.Widgets\MetadataDisplay.cs" />
-    <Compile Include="FSpot.Widgets\PhotoImageView.cs" />
-    <Compile Include="FSpot.Widgets\PreviewPopup.cs" />
-    <Compile Include="FSpot.Widgets\QueryView.cs" />
-    <Compile Include="FSpot.Widgets\RatingMenuItem.cs" />
-    <Compile Include="FSpot.Widgets\Sharpener.cs" />
-    <Compile Include="FSpot.Widgets\Sidebar.cs" />
-    <Compile Include="FSpot.Widgets\SlideShow.cs" />
-    <Compile Include="FSpot.Widgets\SoftFocus.cs" />
-    <Compile Include="FSpot.Widgets\TagEntry.cs" />
-    <Compile Include="FSpot.Widgets\TagMenu.cs" />
-    <Compile Include="FSpot.Widgets\TagView.cs" />
-    <Compile Include="FSpot.Widgets\TrayView.cs" />
-    <Compile Include="FSpot.Widgets\ViewContext.cs" />
+    <Compile Include="Widgets\FolderQueryWidget.cs" />
+    <Compile Include="Widgets\TagSelectionWidget.cs" />
+    <Compile Include="Widgets\CellRendererTextProgress.cs" />
+    <Compile Include="Widgets\EditorPage.cs" />
+    <Compile Include="Widgets\Filmstrip.cs" />
+    <Compile Include="Widgets\FindBar.cs" />
+    <Compile Include="Widgets\FolderTreeModel.cs" />
+    <Compile Include="Widgets\FolderTreePage.cs" />
+    <Compile Include="Widgets\FolderTreeView.cs" />
+    <Compile Include="Widgets\ImageInfo.cs" />
+    <Compile Include="Widgets\InfoBox.cs" />
+    <Compile Include="Widgets\Loupe.cs" />
+    <Compile Include="Widgets\MetadataDisplay.cs" />
+    <Compile Include="Widgets\PhotoImageView.cs" />
+    <Compile Include="Widgets\PreviewPopup.cs" />
+    <Compile Include="Widgets\QueryView.cs" />
+    <Compile Include="Widgets\RatingMenuItem.cs" />
+    <Compile Include="Widgets\Sharpener.cs" />
+    <Compile Include="Widgets\Sidebar.cs" />
+    <Compile Include="Widgets\SlideShow.cs" />
+    <Compile Include="Widgets\SoftFocus.cs" />
+    <Compile Include="Widgets\TagEntry.cs" />
+    <Compile Include="Widgets\TagMenu.cs" />
+    <Compile Include="Widgets\TagView.cs" />
+    <Compile Include="Widgets\TrayView.cs" />
+    <Compile Include="Widgets\ViewContext.cs" />
+    <Compile Include="Widgets\ThumbnailCaptionRenderer.cs" />
+    <Compile Include="Widgets\ThumbnailDateCaptionRenderer.cs" />
+    <Compile Include="Widgets\ThumbnailDecorationRenderer.cs" />
+    <Compile Include="Widgets\ThumbnailFilenameCaptionRenderer.cs" />
+    <Compile Include="Widgets\ThumbnailRatingDecorationRenderer.cs" />
+    <Compile Include="Widgets\ThumbnailTagsCaptionRenderer.cs" />
+    <Compile Include="Widgets\ThumbnailTextCaptionRenderer.cs" />
+    <Compile Include="Widgets\SelectionCollectionGridView.cs" />
+    <Compile Include="Widgets\CollectionCellGridView.cs" />
+    <Compile Include="Widgets\BrowseablePointerGridView.cs" />
+    <Compile Include="Widgets\EditorPageWidget.cs" />
+    <Compile Include="Widgets\LogicEntryCompletion.cs" />
+    <Compile Include="Widgets\CompletionLogic.cs" />
+    <Compile Include="Widgets\StartDragArgs.cs" />
+    <Compile Include="Pinta\ColorBgra.cs" />
     <Compile Include="FSpot\XScreenSaverSlide.cs" />
     <Compile Include="PhotoTagMenu.cs" />
     <Compile Include="PhotoVersionCommands.cs" />
@@ -157,24 +172,10 @@
     <Compile Include="PixbufUtils.cs" />
     <Compile Include="TagPopup.cs" />
     <Compile Include="ThumbnailCommand.cs" />
-    <Compile Include="FSpot.Widgets\ThumbnailCaptionRenderer.cs" />
-    <Compile Include="FSpot.Widgets\ThumbnailDateCaptionRenderer.cs" />
-    <Compile Include="FSpot.Widgets\ThumbnailDecorationRenderer.cs" />
-    <Compile Include="FSpot.Widgets\ThumbnailFilenameCaptionRenderer.cs" />
-    <Compile Include="FSpot.Widgets\ThumbnailRatingDecorationRenderer.cs" />
-    <Compile Include="FSpot.Widgets\ThumbnailTagsCaptionRenderer.cs" />
-    <Compile Include="FSpot.Widgets\ThumbnailTextCaptionRenderer.cs" />
-    <Compile Include="FSpot.Widgets\SelectionCollectionGridView.cs" />
-    <Compile Include="FSpot.Widgets\CollectionCellGridView.cs" />
-    <Compile Include="FSpot.Widgets\BrowseablePointerGridView.cs" />
-    <Compile Include="Pinta\ColorBgra.cs" />
     <Compile Include="Pinta\CairoExtensions.cs" />
     <Compile Include="Pinta\GaussianBlurEffect.cs" />
     <Compile Include="FSpot.ColorAdjustment\AutoStretch.cs" />
     <Compile Include="FSpot.Editors\AutoStretchEditor.cs" />
-    <Compile Include="FSpot.Widgets\EditorPageWidget.cs" />
-    <Compile Include="FSpot.Widgets\LogicEntryCompletion.cs" />
-    <Compile Include="FSpot.Widgets\CompletionLogic.cs" />
     <Compile Include="FSpot.Query\LiteralMenu.cs" />
     <Compile Include="FSpot.Query\TermMenuItem.cs" />
     <Compile Include="FSpot.Query\LiteralBox.cs" />
@@ -188,7 +189,6 @@
     <Compile Include="FSpot.Query\AbstractLiteral.cs" />
     <Compile Include="FSpot.Query\TextLiteral.cs" />
     <Compile Include="FSpot.Query\QueryWidget.cs" />
-    <Compile Include="FSpot.Widgets\StartDragArgs.cs" />
     <Compile Include="FSpot.Import\ImportEvent.cs" />
     <Compile Include="FSpot\ObsoletePhotoQueries.cs" />
     <Compile Include="FSpot.Database\UpdaterUI.cs" />
@@ -399,7 +399,7 @@
   </ItemGroup>
   <ItemGroup>
     <Folder Include="FSpot\" />
-    <Folder Include="FSpot.Widgets\" />
+    <Folder Include="Widgets\" />
     <Folder Include="Pinta\" />
   </ItemGroup>
   <ItemGroup>
