+++ src/Clients/FSpot/Widgets/BrowseablePointerGridView.cs
