+++ src/Clients/FSpot/Widgets/CellRendererTextProgress.cs
