+++ src/Clients/FSpot/Widgets/CollectionCellGridView.cs
