+++ src/Clients/FSpot/Widgets/CompletionLogic.cs
