+++ src/Clients/FSpot/Widgets/EditorPage.cs
