+++ src/Clients/FSpot/Widgets/EditorPageWidget.cs
