+++ src/Clients/FSpot/Widgets/Filmstrip.cs
