+++ src/Clients/FSpot/Widgets/FindBar.cs
