+++ src/Clients/FSpot/Widgets/FolderTreeModel.cs
