+++ src/Clients/FSpot/Widgets/FolderTreePage.cs
