+++ src/Clients/FSpot/Widgets/FolderTreeView.cs
