+++ src/Clients/FSpot/Widgets/ImageInfo.cs
