+++ src/Clients/FSpot/Widgets/InfoBox.cs
