+++ src/Clients/FSpot/Widgets/LogicEntryCompletion.cs
