+++ src/Clients/FSpot/Widgets/Loupe.cs
