+++ src/Clients/FSpot/Widgets/MetadataDisplay.cs
