+++ src/Clients/FSpot/Widgets/PhotoImageView.cs
