+++ src/Clients/FSpot/Widgets/PreviewPopup.cs
