+++ src/Clients/FSpot/Widgets/QueryView.cs
