+++ src/Clients/FSpot/Widgets/RatingMenuItem.cs
