+++ src/Clients/FSpot/Widgets/SelectionCollectionGridView.cs
