+++ src/Clients/FSpot/Widgets/Sharpener.cs
