+++ src/Clients/FSpot/Widgets/Sidebar.cs
