+++ src/Clients/FSpot/Widgets/SlideShow.cs
