+++ src/Clients/FSpot/Widgets/SoftFocus.cs
