+++ src/Clients/FSpot/Widgets/StartDragArgs.cs
