+++ src/Clients/FSpot/Widgets/TagEntry.cs
