+++ src/Clients/FSpot/Widgets/TagMenu.cs
