+++ src/Clients/FSpot/Widgets/TagSelectionWidget.cs
