+++ src/Clients/FSpot/Widgets/TagView.cs
