+++ src/Clients/FSpot/Widgets/ThumbnailCaptionRenderer.cs
