+++ src/Clients/FSpot/Widgets/ThumbnailDateCaptionRenderer.cs
