+++ src/Clients/FSpot/Widgets/ThumbnailDecorationRenderer.cs
