+++ src/Clients/FSpot/Widgets/ThumbnailFilenameCaptionRenderer.cs
