+++ src/Clients/FSpot/Widgets/ThumbnailRatingDecorationRenderer.cs
