+++ src/Clients/FSpot/Widgets/ThumbnailTagsCaptionRenderer.cs
