+++ src/Clients/FSpot/Widgets/ThumbnailTextCaptionRenderer.cs
