+++ src/Clients/FSpot/Widgets/TrayView.cs
