+++ src/Clients/FSpot/Widgets/ViewContext.cs
